Extract camera content fit logic
